Refactored to fit a standard template.
